requirements.txt
boto3==1.5.21
botocore==1.8.35
docutils==0.14
futures==3.2.0
jmespath==0.9.3
python-dateutil==2.6.1
PyYAML==3.12
s3transfer==0.1.12
six==1.11.0
